Separate logic from view
